Finalizando o Projeto Python HAVI, adicionando Layout ao form de Contato e começando Demo e Solu
