--- Front-End_WEB GS/Contato.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="css/contato.css">
-    <title>Contato</title>
-</head>
-<body>
-    <div class="container">
-
-        <div class="box_login">
-            <h1 class="title">Entre em contato </h1>
-        
-            <form action="#" mothod="post"> 
-                <input type="text" name="nome" id="nome" placeholder="Nome">
-
-                <input type="email" name="email" id="email" placeholder="E-mail">
-
-                <button type="submit">Entrar</button>
-                <div class="divider"></div>
-             
-            </form>
-        
-        </div>
-    
-    </div>
-    
-</body>
-</html>
