--- Front-End_WEB GS/Python GS/README.txt
@@ -1,4 +1,0 @@
-Autor: Artur Rodrigues Loreto ;
-RM: 552851 ;
-Curso: Enhenharia de Software, 1S23.
-
