--- Front-End_WEB GS/css/contato.css
@@ -1,101 +1,0 @@
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-.container{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 100vh;
-    background-image: url('../img/pexels-los-muertos-crew-8460049.jpg');
-    background-position: center center;
-    background-size: cover;
-    background-repeat: no-repeat;
-}
-
-.box_login{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 400px;
-    height: 460px;
-    padding: 50px;
-    border-radius: 10px;
-    text-align: center;
-    backdrop-filter: blur(16px);
-    -webkit-backdrop-filter: blur(16px);
-    border: 2px solid rgba(255, 255, 255, .2);
-}
-
-.box_login .title{
-    color: #ccc;
-    font-size: 24px;
-    text-shadow: 0 0 10px rgba(0,0,0,.5);
-
-}
-
-.box_login form{
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    width: 100%;
-    margin-top: 32px;
-}
-
-.box_login input{
-    outline: none;
-    background-color: transparent;
-    color: #ccc;
-    width: 100%;
-    height: 48px;
-    border: 2px solid rgba(255, 255, 255, .2);
-    padding-left: 16px;
-    font-size: 14px;
-    margin-bottom: 16px;
-    border-radius: 16px;
-}
-
-
-
-.box_login input:focus{
-    border: 2px solid rgba(255, 255, 255, .3)
-}
-
-.box_login input::placeholder{
-    color: #ccc;
-}
-
-.box_login button{
-    border: none;
-    outline: none;
-    width: 100%;
-    height: 48px;
-    background-color: #1e90ff;
-    color: #fff;
-    border-radius:  10px;
-    cursor: pointer;
-}
-
-.box_login .divider{
-    display: block;
-    width: 80%;
-    height: 1px;
-    background-color: #fff;
-    margin-top:32px;
-    margin-bottom: 32px;
-    opacity: .2;
-}
-
-.box_login a{
-    font-size: 16px;
-    color: #fff;
-    text-decoration: none;
-}
-
-.box_login a:hover{
-    text-decoration: underline;
-}
