--- Front-End_WEB GS/css/estilo.css
@@ -1,226 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Inconsolata&family=Playfair+Display:ital@1&family=Raleway:wght@200&family=Roboto:ital,wght@1,300&display=swap');
-*{
-
-    margin: 0;
-    padding: 0;
-    /*font-family: 'Inconsolata', monospace;
-    font-family: 'Playfair Display', serif;
-    font-family: 'Raleway', sans-serif;*/
-    font-family:'Poppins', sans-serif;
-}
-
-
-
-body{
-    overflow-x: hidden;
-    background: linear-gradient(to bottom, snow, rgb(240, 221, 185));
-    background-repeat:no-repeat;
-    background-size:cover;
-    height: 100vh;
-}
-
-.container-sec{
-    justify-content: center;
-    text-align: right;
-    left: 2%;
-}
-
-.nav-icon{
-    display:inline-block;
-    height: 10vh;
-    transition: all;
-    bottom: 6rem;
-    transition: background 0.2s linear;
-    position: relative;
-}
-
-body.dark{
-    background-color: #232c35;
-}
-
-#chk{
-    opacity: 0;
-    position: relative;
-
-}
-
-
-.label{
-    background-color: #111;
-    border-radius: 60px;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 5px;
-    position: relative;
-    height: 15px;
-    width: 48px;
-    transform: scale(1);
-    top: 20px;
-    right: 5rem;
-}
-
-.label .ball{
-    background-color: #fff;
-    border-radius: 55%;
-    position: absolute;
-    top: 1.4px;
-    left: 2px;
-    height: 20px;
-    width: 20px;
-    transform: translateX(0px);
-    transition: transform 0.2s linear;
-}
-
-#chk:checked + .label .ball{
-    transform: translateX(33px);
-}
-
-.fa-moon{
-    color: #f1c40f;
-}
-
-.fa-sun{
-    color: #f39c12;
-}
-
-#img-HapNotre{ /* Logo */
-    height: 2.8rem;
-    position:absolute;
-    display: flex;
-    cursor: pointer;
-    margin-top: 10px;
-    left: 2%;
-    padding: 10px;
-}
-
-.container{
-    display:inline-table;
-    background-image:linear-gradient(to right, #ffc53cd8,#ff312e);
-    /*box-shadow: 0 3px lightgrey;*/
-    box-shadow:  0px 0px 5px 1px;
-    width: 100%;
-    justify-content:center;
-    height:9.5vh ;
-    text-align: center;
-    opacity: 100%;
-}
-
-
-
-.btn-nav{ /* Elementos da vavega??o */
-    display: flex;
-    gap: 1px;
-    position: relative;
-    list-style: none;
-    color: #57575c;
-    align-items: center;
-    font-size: 17px;
-    justify-content: center;
-}
-
-.btn-nav li a{
-    padding:34px;
-    text-decoration: none;
-    color: rgb(68, 68, 68);
-    display: inline-block;
-    position: relative;
-    letter-spacing: 0.5px;
-}
-
-
-.btn-nav li a:after{
-    content: "";
-    position: absolute;
-    height: 3px;
-    width: 0;
-    background-color:  #2c6cbc;
-    color: snow;
-    left: 0;
-    bottom:22px;
-    transition: 0.5s;
-    
-}
-
-.btn-nav li a:hover{
-    color:#ffffff;
-    
-}
-
-.btn-nav li a:hover:after{
-    width: 100%;
-}
-
-#parag_ini{
-    font-size: 30px;
-    text-align: center;
-    color: whitesmoke;
-}
-
-
-.rodape{
-    border: 0ch solid #484d52;
-    border-radius: 20px 1px 1px 1px;
-    background:#484d52;
-    position: absolute;
-    width: 100%;
-    height: 13rem;                          /*Atribuindo altura do rodapé*/
-    bottom: 0;                              /*Faz o rodapé ficar em baixo*/
-    font-size: 15px;
-    padding-top:-20px;
-    box-shadow:0px 0px 6px 3px ;
-}
-
-
-
-.p_footer{
-    position: absolute;
-    color: #fff;
-    margin-top: 20px;
-    bottom: 20px;
-    letter-spacing: 1.5px;
-    width: 100%;
-    text-align: center;
-    position:absolute;
-}
-
-#Inst{
-    height: 32px;
-}
-
-#Linkd{
-    display: inline;
-    position: relative;
-    height: 35px;
-}
-
-#Git{
-    height: 35px;
-}
-
-.img-midia{
-    display:inline-flex;
-    position: relative;
-    width: 100%;
-    height: 100%;
-    justify-content: center;
-    align-items: center;
-}
-
-hr{
-
-}
-
-
-
-/*
-    width: 10px;
-    display: inline;
-    position: relative;
-    bottom: 15px;
-    letter-spacing: 30px;
-    text-align: center;
-    justify-content: center;
-    align-items: center;
-    justify-self: center;
