--- Front-End_WEB GS/home.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="css/estilo.css">
-    <title>Hapvida</title>
-</head>
-<body>
-
-<h1></h1>
-    <header>
-        <section class="container-sec">
-            <div >
-                <div>
-                    <img id="img-HapNotre" src="img/logo-hapvida-NotreDame-Intermedica.png" alt="">
-                </div>
-                <nav class="container">
-                    <div id="btn-nav">
-                        <ul class="btn-nav">
-                            <li id="home"><a href="#home"><p> Home</p></a></li>
-                            <li id="problema"><a href="#Problema"><p> Problema & Solução</p></a></li>
-                            <li id="demonstracao"><a href="#Demonstração de Solução"><p> Demonstração e Solução</p></a></li>
-                            <li id="contato"><a href="#Contato"><p> Contato</p></a></li>
-                        </ul>
-                    </div>
-                </nav>
-                <div class="nav-icon">
-                    <input type="checkbox" class="checkbox" id="chk">
-                    <label class="label" for="chk">
-                        <i class="fas fa-sun"></i>
-                        <i class="fas fa-moon"></i>
-                        <div class="ball"></div>
-                    </label>
-                </div>
-
-            </div>
-            
-        </section>
-    </header>
-
-    <div id="parag_ini">
-        <p >Com tanto crescimento da Hapvida nos últimos 7 anos segundo a bolsa, é evidente a necessidade de organizção de dados.</p>
-        <p>Com esse intuito foi criado o SiHop uma aplicação web de organização, cadastro e controle de entrada e saída de pacientes.</p>
-    </div>
-
-
-    <footer class="rodape">
-        <div class="img-midia">
-            <a href="" target="_blank" ><img id="Inst"src="img/iconInstagram.png" alt="Icone Instagram"></a>
-            <a href="https://www.linkedin.com/in/artur-loreto-82bb71267/" target="_blank"><img id="Linkd" src="img/iconLinkedin.png" alt="Icone Linkedin"></a>
-            <a href="https://github.com/Loreto1306" target="_blank"><img id="Git" src="img/iconGitHub.png" alt="Icone GitHub"></a>
-            
-        </div>
-
-        <div class="p_footer">
-            <p>© 2023 Artur Loreto Todos os direitos reservados</p>
-        </div>
-
-    </footer>
-
-
-<script src="https://kit.fontawesome.com/998c60ef77.js" crossorigin="anonymous"></script>
-<script src="script/script.js"></script>
-</body>
-</html>
