--- Front-End_WEB GS/script/script.js
@@ -1,5 +1,0 @@
-const chk = document.getElementById('chk')
-
-chk.addEventListener('change', ()=> {
-    document.body.classList.toggle('dark')
-})
