+++ Python GS/README.txt
@@ -1,0 +1,4 @@
+Autor: Artur Rodrigues Loreto ;
+RM: 552851 ;
+Curso: Enhenharia de Software, 1S23.
+
