+++ css/contato.css
@@ -1,0 +1,330 @@
+@import url('https://fonts.googleapis.com/css2?family=Inconsolata&family=Playfair+Display:ital@1&family=Raleway:wght@200&family=Roboto:ital,wght@1,300&display=swap');
+*{
+
+    margin: 0;
+    padding: 0;
+    /*font-family: 'Inconsolata', monospace;
+    font-family: 'Playfair Display', serif;
+    font-family: 'Raleway', sans-serif;*/
+    font-family:'Poppins', sans-serif;
+
+}
+
+
+
+body{
+    overflow-x: hidden;
+    overflow-y: hidden;
+    background-image: url('../img/photo-1516574187841-cb9cc2ca948b.avif');
+    background-position: center center;
+    background-size: cover;
+    background-repeat: no-repeat;
+    height: 100vh;
+}
+
+.container-sec{
+    justify-content: center;
+    text-align: right;
+    left: 2%;
+}
+
+.nav-icon{
+    display:inline-block;
+
+    transition: all;
+    bottom: 6rem;
+    transition: background 0.2s linear;
+    position: relative;
+}
+
+body.dark{
+    background-color: #232c35;
+}
+
+#chk{
+    opacity: 0;
+    position: relative;
+
+}
+
+
+.label{
+    background-color: #111;
+    border-radius: 60px;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 5px;
+    position: relative;
+    height: 15px;
+    width: 48px;
+    transform: scale(1);
+    top: 20px;
+    right: 5rem;
+}
+
+.label .ball{
+    background-color: #fff;
+    border-radius: 55%;
+    position: absolute;
+    top: 1.4px;
+    left: 2px;
+    height: 20px;
+    width: 20px;
+    transform: translateX(0px);
+    transition: transform 0.2s linear;
+}
+
+#chk:checked + .label .ball{
+    transform: translateX(33px);
+}
+
+.fa-moon{
+    color: #f1c40f;
+}
+
+.fa-sun{
+    color: #f39c12;
+}
+
+#img-HapNotre{ /* Logo */
+    height: 2.8rem;
+    position:absolute;
+    display: flex;
+    cursor: pointer;
+    margin-top: 10px;
+    left: 2%;
+    padding: 10px;
+}
+
+.container{
+    display:inline-table;
+    background-image:linear-gradient(to right, #ffc53cd8,#ff312e);
+    /*box-shadow: 0 3px lightgrey;*/
+    box-shadow:  0px 0px 5px 1px;
+    width: 100%;
+    justify-content:center;
+    height:9.5vh ;
+    text-align: center;
+    opacity: 100%;
+}
+
+
+
+.btn-nav{ /* Elementos da vavega??o */
+    display: flex;
+    gap: 1px;
+    position: relative;
+    list-style: none;
+    color: #57575c;
+    align-items: center;
+    font-size: 17px;
+    justify-content: center;
+}
+
+.btn-nav li a{
+    padding:34px;
+    text-decoration: none;
+    color: rgb(68, 68, 68);
+    display: inline-block;
+    position: relative;
+    letter-spacing: 0.5px;
+}
+
+
+.btn-nav li a:after{
+    content: "";
+    position: absolute;
+    height: 3px;
+    width: 0;
+    background-color:  #2c6cbc;
+    color: snow;
+    left: 0;
+    bottom:22px;
+    transition: 0.5s;
+    
+}
+
+.btn-nav li a:hover{
+    color:#ffffff;
+    
+}
+
+.btn-nav li a:hover:after{
+    width: 100%;
+}
+
+#parag_ini{
+    font-size: 30px;
+    text-align: center;
+    color: whitesmoke;
+}
+
+
+.rodape{
+    border: 0ch solid #484d52;
+    border-radius: 20px 1px 1px 1px;
+    background:#484d52;
+    position: absolute;
+    width: 100%;
+    height: 14vh;                          /*Atribuindo altura do rodapé*/
+    bottom: 0;                              /*Faz o rodapé ficar em baixo*/
+    font-size: 15px;
+    padding-top:-20px;
+    box-shadow:0px 0px 6px 3px ;
+}
+
+
+
+.p_footer{
+    position: absolute;
+    color: #fff;
+    margin-top: 20px;
+    bottom: 20px;
+    letter-spacing: 1.5px;
+    width: 100%;
+    text-align: center;
+    position:absolute;
+}
+
+#Inst{
+    height: 32px;
+}
+
+#Linkd{
+    display: inline;
+    position: relative;
+    height: 35px;
+}
+
+#Git{
+    height: 35px;
+}
+
+.img-midia{
+    display:inline-flex;
+    position: relative;
+    width: 100%;
+    height: 100%;
+    justify-content: center;
+    align-items: center;
+}
+
+
+#container{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 80vh;
+}
+
+
+
+.box_login{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 400px;
+    height:450px;
+    margin-bottom: 5%;
+    padding: 60px;
+    border-radius: 10px;
+    text-align: center;
+    backdrop-filter: blur(20px);
+    -webkit-backdrop-filter: blur(20px);
+    border: 2px solid rgba(255, 255, 255, .2);
+}
+
+.box_login .title{
+    color: #ffffff;
+    font-size: 24px;
+    text-shadow: 0 0 10px rgba(0,0,0,.5);
+
+}
+
+.box_login form{
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    width: 100%;
+    margin-top: 32px;
+}
+
+.box_login input{
+    margin: 5px;
+    outline: none;
+    background-color: transparent;
+    color: #ffffff;
+    width: 100%;
+    height: 48px;
+    border: 2px solid rgba(255, 255, 255, .2);
+    padding-left: 16px;
+    font-size: 14px;
+    margin-bottom: 16px;
+    border-radius: 16px;
+}
+
+
+
+.box_login input:focus{
+    border: 2px solid rgba(255, 255, 255, .3)
+}
+
+.box_login input::placeholder{
+    color: #ffffff;
+}
+
+.box_login button{
+    margin-top: 20px;
+    border: none;
+    outline: none;
+    width:70%;
+    height: 40px;
+    background-color: #1e90ff;
+    color: #fff;
+    border-radius:  10px;
+    cursor: pointer;
+}
+
+.box_login button:hover{background-color: #147be2;}
+
+
+.box_login .divider{
+    display: block;
+    width: 80%;
+    height: 1px;
+    background-color: #fff;
+    margin-top:70px;
+    margin-bottom: 40px;
+    opacity: .2;
+}
+
+.box_login a{
+    font-size: 16px;
+    color: #fff;
+    text-decoration: none;
+}
+
+.box_login a:hover{
+    text-decoration: underline;
+}
+
+hr{
+
+}
+
+
+
+/*
+    width: 10px;
+    display: inline;
+    position: relative;
+    bottom: 15px;
+    letter-spacing: 30px;
+    text-align: center;
+    justify-content: center;
+    align-items: center;
+    justify-self: center;
