+++ script/script.js
@@ -1,0 +1,5 @@
+const chk = document.getElementById('chk')
+
+chk.addEventListener('change', ()=> {
+    document.body.classList.toggle('dark')
+})
